feat: テナントキャッシュ機能の追加と設定の改善
- config.rs: テナントキャッシュ設定を追加し、デフォルト値を設定
- kafka_producer.rs: 冪等プロデューサーを有効化し、メッセージの重複送信を防止
- keycloak_admin.rs: HTTPクライアントのタイムアウトを30秒に設定
- mod.rs: cacheモジュールを追加
- saga_client.rs: HTTPクライアントのタイムアウトを30秒に設定
- startup.rs: テナントキャッシュ設定をConfigから取得し、リポジトリをキャッシュで包む
- integration_test.rs: テスト用HTTPクライアントのタイムアウトを10秒に設定
- config.yaml: データベース設定を有効化し、vault_dbに変更
- encryption.rs: 環境変数からの鍵の読み込み時に本番環境でのゼロ鍵使用を拒否
- kafka_producer.rs: 冪等プロデューサーを有効化し、メッセージの重複送信を防止
- config.docker.yaml: スケジューラーの内部エンドポイントを修正
- workflow/config.yaml: データベース設定を有効化し、workflow_dbに変更
- dto.rs: エラーレスポンス生成のヘルパーを改善
- instance_handler.rs: レスポンスDTOを直接Jsonとして返すように修正
- workflow_handler.rs: レスポンスDTOを直接Jsonとして返すように修正
- notification_request_producer.rs: 冪等プロデューサーを有効化し、メッセージの重複送信を防止
- cargo-audit.sh: 脆弱性監査スクリプトを改善し、ビジネス層とサービス層のクレートをスキャン
- dart-outdated.sh: outdatedパッケージ検出時に非ゼロ終了コードを返すように修正

+++ regions/system/server/rust/graphql-gateway/src/infrastructure/grpc/auth_client.rs
@@ -184,7 +184,7 @@ impl AuthGrpcClient {
         user_id: Option<&str>,
         event_type: Option<&str>,
         result: Option<&str>,
-    ) -> anyhow::Result<(Vec<AuditLog>, i32, bool)> {
+    ) -> anyhow::Result<(Vec<AuditLog>, i64, bool)> {
         let request = tonic::Request::new(proto::k1s0::system::auth::v1::SearchAuditLogsRequest {
             pagination: Some(proto::k1s0::system::common::v1::Pagination {
                 page: page.unwrap_or(1),
